+++ .github/workflows/build.yml
@@ -594,9 +594,13 @@ jobs:
         tar -xvf r2-static-latest.tar.xz -C r2-static-latest --strip-components=1
     - name: ABI compatibility check
       run: |
-        echo '[suppress_type]' > abidiff.supressions.conf
-        echo 'name_regexp = ^cs_' >> abidiff.supressions.conf
-        for lib in r2-static-latest/usr/lib/libr_*.so; do abidiff --suppr abidiff.supressions.conf ${lib} r2-static/usr/lib/$(basename ${lib}); done
+        echo '[suppress_function]' > abidiff.conf
+        echo 'name_not_regexp = ^r_.*' >> abidiff.conf
+        echo '[suppress_variable]' >> abidiff.conf
+        echo 'name_not_regexp = ^r_.*' >> abidiff.conf
+        for lib in r2-static-latest/usr/lib/libr_*.so; do
+          echo $lib ; abidiff --suppr abidiff.conf ${lib} r2-static/usr/lib/$(basename ${lib});
+        done
 
 
   # Release creation
