+++ libr/include/r_arch.h
@@ -330,6 +330,7 @@ extern RArchPlugin r_arch_plugin_h8300;
 extern RArchPlugin r_arch_plugin_bf;
 extern RArchPlugin r_arch_plugin_sparc_gnu;
 extern RArchPlugin r_arch_plugin_hppa_gnu;
+extern RArchPlugin r_arch_plugin_s390_cs;
 extern RArchPlugin r_arch_plugin_s390_gnu;
 extern RArchPlugin r_arch_plugin_m68k_gnu;
 extern RArchPlugin r_arch_plugin_ppc_gnu;
